requirements.txt
-i https://pypi.org/simple
asyncpg==0.20.1
bcrypt==3.1.7
cffi==1.14.0
click==7.0
dataclasses==0.7 ; python_version < '3.7'
fastapi==0.49.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
passlib==1.7.2
pycparser==2.19
pydantic==1.4
pyjwt==1.7.1
python-multipart==0.0.5
six==1.14.0
starlette==0.12.9
uvicorn==0.11.3
uvloop==0.14.0 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1
